+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2026.2.4] — 2026-06-13
 
-
+### Added
+- **Model Showcase** - Customizable background per model, with a nice opaque UI.
 
 ### Fixed
 - **Model Dropdown** - fixed a bug where if an input is too large, model dropdown would clip and not allow users to select models.